chore: finish renaming consolidated and move tests
Ref. metriport/metriport-internal#870
