+++ packages/api/src/external/commonwell/document/document-query.ts
@@ -15,10 +15,13 @@ import {
   S3Info,
   uploadStream,
 } from "../../../command/medical/document/document-query-storage-info";
+import {
+  MAPIWebhookStatus,
+  processPatientDocumentRequest,
+} from "../../../command/medical/document/document-webhook";
 import { appendDocQueryProgress } from "../../../command/medical/patient/append-doc-query-progress";
 import { getPatientOrFail } from "../../../command/medical/patient/get-patient";
 import { ApiTypes, reportUsage } from "../../../command/usage/report-usage";
-import { MAPIWebhookStatus, processPatientDocumentRequest } from "../../../command/webhook/medical";
 import ConversionError from "../../../errors/conversion-error";
 import MetriportError from "../../../errors/metriport-error";
 import { MedicalDataSource } from "../../../external";
